+++ app/modules/kernel/src/Controller/ControllerResolver.php
@@ -81,7 +81,7 @@ class ControllerResolver
     protected function doGetArguments(Request $request, $controller, array $parameters)
     {
         $attributes = $request->attributes->all();
-        $arguments = array();
+        $arguments = [];
         foreach ($parameters as $param) {
             if (array_key_exists($param->name, $attributes)) {
                 $arguments[] = $attributes[$param->name];
@@ -124,7 +124,7 @@ class ControllerResolver
             throw new \InvalidArgumentException(sprintf('Class "%s" does not exist.', $class));
         }
 
-        return array($this->instantiateController($class), $method);
+        return [$this->instantiateController($class), $method];
     }
 
     /**