+++ app/modules/view/index.php
@@ -1,6 +1,7 @@
 <?php
 
 use Pagekit\Event\PrefixEventDispatcher;
+use Pagekit\Twig\TwigEngine;
 use Pagekit\View\Asset\AssetFactory;
 use Pagekit\View\Asset\AssetManager;
 use Pagekit\View\Helper\DataHelper;
@@ -18,19 +19,35 @@ use Pagekit\View\Loader\FilesystemLoader;
 use Pagekit\View\PhpEngine;
 use Pagekit\View\View;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Templating\TemplateNameParser;
 
 return [
 
     'name' => 'view',
+
+    'include' => 'modules/*/index.php',
+
+    'require' => [
+
+        'view/twig'
+
+    ],
 
     'main' => function ($app) {
 
         $app['view'] = function ($app) {
 
             $view = new View(new PrefixEventDispatcher('view.', $app['events']));
+
             $view->addEngine(new PhpEngine(null, isset($app['locator']) ? new FilesystemLoader($app['locator']) : null));
+
+            if (isset($app['twig'])) {
+                $view->addEngine(new TwigEngine($app['twig'], new TemplateNameParser()));
+            }
+
             $view->addGlobal('app', $app);
             $view->addGlobal('view', $view);
+
             $view->addHelpers([
                 new DataHelper(),
                 new DeferredHelper($app['events']),