+++ app/system/modules/user/src/Controller/UserController.php
@@ -23,7 +23,7 @@ class UserController
         return [
             '$view' => [
                 'title' => __('Users'),
-                'name' => 'system/user:views/admin/user-index.php'
+                'name' => 'system/user/admin/user-index.php'
             ],
             '$data' => [
                 'statuses' => User::getStatuses(),
@@ -54,7 +54,7 @@ class UserController
         return [
             '$view' => [
                 'title' => $id ? __('Edit User') : __('Add User'),
-                'name' => 'system/user:views/admin/user-edit.php'
+                'name' => 'system/user/admin/user-edit.php'
             ],
             '$data' => [
                 'user' => $user,
@@ -76,7 +76,7 @@ class UserController
         return [
             '$view' => [
                 'title' => __('Permissions'),
-                'name' => 'system/user:views/admin/permission-index.php'
+                'name' => 'system/user/admin/permission-index.php'
             ],
             '$data' => [
                 'permissions' => App::module('system/user')->getPermissions(),
@@ -94,7 +94,7 @@ class UserController
         return [
             '$view' => [
                 'title' => __('Roles'),
-                'name' => 'system/user:views/admin/role-index.php'
+                'name' => 'system/user/admin/role-index.php'
             ],
             '$config' => [
                 'role' => $id
@@ -114,7 +114,7 @@ class UserController
         return [
             '$view' => [
                 'title' => __('User Settings'),
-                'name' => 'system/user:views/admin/settings.php'
+                'name' => 'system/user/admin/settings.php'
             ],
             '$data' => [
                 'config' => App::module('system/user')->config()